+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 
 ### Fixed
 * Made correction to properly read boolean 'false' from settings.
+* Changes to WSAD controller for more reliable key press captures.
 
 ## 0.0.9 - 2023-03-26